requirements.txt
pandas
matplotlib
seaborn
plotly
mysql-connector-python
streamlit